requirements.txt
# ─────────────────────────────────────────────
# FinQuery AI — Python Dependencies
# ─────────────────────────────────────────────

# Core Framework
django>=5.0,<7.0
djangorestframework>=3.15
django-cors-headers>=4.3

# Async Task Queue
celery>=5.4
django-celery-results>=2.5
redis>=5.0

# Database & ORM
sqlalchemy>=2.0

# LLM Integration
ollama>=0.4

# Data Processing (used by helpers/)
pandas>=2.2
faker>=25.0

# Production WSGI Server
gunicorn>=22.0
